requirements.txt
django
djangorestframework
psycopg2
requests
pytz
tzlocal
